requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
iniconfig==1.1.1
mccabe==0.6.1
python-dotenv==0.19.0
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
schedule==1.1.0
six==1.16.0
snowballstemmer==2.2.0
toml==0.10.2
tornado==6.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
redis==5.0.4
peewee==3.17.5
psycopg2==2.9.9

